+++ exporter/opentelemetry-exporter-otlp-proto-http/tests/metrics/test_otlp_metrics_exporter.py
@@ -34,10 +34,14 @@ from opentelemetry.exporter.otlp.proto.http.metric_exporter import (
 )
 from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_OTLP_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_CLIENT_KEY,
     OTEL_EXPORTER_OTLP_COMPRESSION,
     OTEL_EXPORTER_OTLP_ENDPOINT,
     OTEL_EXPORTER_OTLP_HEADERS,
     OTEL_EXPORTER_OTLP_METRICS_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_METRICS_CLIENT_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_METRICS_CLIENT_KEY,
     OTEL_EXPORTER_OTLP_METRICS_COMPRESSION,
     OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION,
     OTEL_EXPORTER_OTLP_METRICS_ENDPOINT,
@@ -73,6 +77,8 @@ from opentelemetry.test.metrictestutil import _generate_sum
 
 OS_ENV_ENDPOINT = "os.env.base"
 OS_ENV_CERTIFICATE = "os/env/base.crt"
+OS_ENV_CLIENT_CERTIFICATE = "os/env/client-cert.pem"
+OS_ENV_CLIENT_KEY = "os/env/client-key.pem"
 OS_ENV_HEADERS = "envHeader1=val1,envHeader2=val2"
 OS_ENV_TIMEOUT = "30"
 
@@ -114,6 +120,8 @@ class TestOTLPMetricExporter(TestCase):
             exporter._endpoint, DEFAULT_ENDPOINT + DEFAULT_METRICS_EXPORT_PATH
         )
         self.assertEqual(exporter._certificate_file, True)
+        self.assertEqual(exporter._client_certificate_file, None)
+        self.assertEqual(exporter._client_key_file, None)
         self.assertEqual(exporter._timeout, DEFAULT_TIMEOUT)
         self.assertIs(exporter._compression, DEFAULT_COMPRESSION)
         self.assertEqual(exporter._headers, {})
@@ -123,11 +131,15 @@ class TestOTLPMetricExporter(TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_ENDPOINT: OS_ENV_ENDPOINT,
             OTEL_EXPORTER_OTLP_HEADERS: OS_ENV_HEADERS,
             OTEL_EXPORTER_OTLP_TIMEOUT: OS_ENV_TIMEOUT,
             OTEL_EXPORTER_OTLP_METRICS_CERTIFICATE: "metrics/certificate.env",
+            OTEL_EXPORTER_OTLP_METRICS_CLIENT_CERTIFICATE: "metrics/client-cert.pem",
+            OTEL_EXPORTER_OTLP_METRICS_CLIENT_KEY: "metrics/client-key.pem",
             OTEL_EXPORTER_OTLP_METRICS_COMPRESSION: Compression.Deflate.value,
             OTEL_EXPORTER_OTLP_METRICS_ENDPOINT: "https://metrics.endpoint.env",
             OTEL_EXPORTER_OTLP_METRICS_HEADERS: "metricsEnv1=val1,metricsEnv2=val2,metricEnv3===val3==",
@@ -139,6 +151,10 @@ class TestOTLPMetricExporter(TestCase):
 
         self.assertEqual(exporter._endpoint, "https://metrics.endpoint.env")
         self.assertEqual(exporter._certificate_file, "metrics/certificate.env")
+        self.assertEqual(
+            exporter._client_certificate_file, "metrics/client-cert.pem"
+        )
+        self.assertEqual(exporter._client_key_file, "metrics/client-key.pem")
         self.assertEqual(exporter._timeout, 40)
         self.assertIs(exporter._compression, Compression.Deflate)
         self.assertEqual(
@@ -155,6 +171,8 @@ class TestOTLPMetricExporter(TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_ENDPOINT: OS_ENV_ENDPOINT,
             OTEL_EXPORTER_OTLP_METRICS_ENDPOINT: "https://metrics.endpoint.env",
@@ -166,6 +184,8 @@ class TestOTLPMetricExporter(TestCase):
         exporter = OTLPMetricExporter(
             endpoint="example.com/1234",
             certificate_file="path/to/service.crt",
+            client_key_file="path/to/client-key.pem",
+            client_certificate_file="path/to/client-cert.pem",
             headers={"testHeader1": "value1", "testHeader2": "value2"},
             timeout=20,
             compression=Compression.NoCompression,
@@ -174,6 +194,10 @@ class TestOTLPMetricExporter(TestCase):
 
         self.assertEqual(exporter._endpoint, "example.com/1234")
         self.assertEqual(exporter._certificate_file, "path/to/service.crt")
+        self.assertEqual(
+            exporter._client_certificate_file, "path/to/client-cert.pem"
+        )
+        self.assertEqual(exporter._client_key_file, "path/to/client-key.pem")
         self.assertEqual(exporter._timeout, 20)
         self.assertIs(exporter._compression, Compression.NoCompression)
         self.assertEqual(
@@ -186,6 +210,8 @@ class TestOTLPMetricExporter(TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_HEADERS: OS_ENV_HEADERS,
             OTEL_EXPORTER_OTLP_TIMEOUT: OS_ENV_TIMEOUT,
@@ -196,6 +222,10 @@ class TestOTLPMetricExporter(TestCase):
         exporter = OTLPMetricExporter()
 
         self.assertEqual(exporter._certificate_file, OS_ENV_CERTIFICATE)
+        self.assertEqual(
+            exporter._client_certificate_file, OS_ENV_CLIENT_CERTIFICATE
+        )
+        self.assertEqual(exporter._client_key_file, OS_ENV_CLIENT_KEY)
         self.assertEqual(exporter._timeout, int(OS_ENV_TIMEOUT))
         self.assertIs(exporter._compression, Compression.Gzip)
         self.assertEqual(
@@ -244,7 +274,8 @@ class TestOTLPMetricExporter(TestCase):
                 (
                     "Header format invalid! Header values in environment "
                     "variables must be URL encoded per the OpenTelemetry "
-                    "Protocol Exporter specification: missingValue"
+                    "Protocol Exporter specification or a comma separated "
+                    "list of name=value occurrences: missingValue"
                 ),
             )
 
@@ -294,6 +325,7 @@ class TestOTLPMetricExporter(TestCase):
             data=serialized_data.SerializeToString(),
             verify=exporter._certificate_file,
             timeout=exporter._timeout,
+            cert=exporter._client_cert,
         )
 
     @activate