+++ exporter/opentelemetry-exporter-otlp-proto-http/tests/test_proto_span_exporter.py
@@ -29,11 +29,15 @@ from opentelemetry.exporter.otlp.proto.http.trace_exporter import (
 from opentelemetry.exporter.otlp.proto.http.version import __version__
 from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_OTLP_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_CLIENT_KEY,
     OTEL_EXPORTER_OTLP_COMPRESSION,
     OTEL_EXPORTER_OTLP_ENDPOINT,
     OTEL_EXPORTER_OTLP_HEADERS,
     OTEL_EXPORTER_OTLP_TIMEOUT,
     OTEL_EXPORTER_OTLP_TRACES_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_TRACES_CLIENT_CERTIFICATE,
+    OTEL_EXPORTER_OTLP_TRACES_CLIENT_KEY,
     OTEL_EXPORTER_OTLP_TRACES_COMPRESSION,
     OTEL_EXPORTER_OTLP_TRACES_ENDPOINT,
     OTEL_EXPORTER_OTLP_TRACES_HEADERS,
@@ -44,6 +48,8 @@ from opentelemetry.sdk.trace.export import SpanExportResult
 
 OS_ENV_ENDPOINT = "os.env.base"
 OS_ENV_CERTIFICATE = "os/env/base.crt"
+OS_ENV_CLIENT_CERTIFICATE = "os/env/client-cert.pem"
+OS_ENV_CLIENT_KEY = "os/env/client-key.pem"
 OS_ENV_HEADERS = "envHeader1=val1,envHeader2=val2"
 OS_ENV_TIMEOUT = "30"
 
@@ -58,6 +64,8 @@ class TestOTLPSpanExporter(unittest.TestCase):
             exporter._endpoint, DEFAULT_ENDPOINT + DEFAULT_TRACES_EXPORT_PATH
         )
         self.assertEqual(exporter._certificate_file, True)
+        self.assertEqual(exporter._client_certificate_file, None)
+        self.assertEqual(exporter._client_key_file, None)
         self.assertEqual(exporter._timeout, DEFAULT_TIMEOUT)
         self.assertIs(exporter._compression, DEFAULT_COMPRESSION)
         self.assertEqual(exporter._headers, {})
@@ -76,11 +84,15 @@ class TestOTLPSpanExporter(unittest.TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_ENDPOINT: OS_ENV_ENDPOINT,
             OTEL_EXPORTER_OTLP_HEADERS: OS_ENV_HEADERS,
             OTEL_EXPORTER_OTLP_TIMEOUT: OS_ENV_TIMEOUT,
             OTEL_EXPORTER_OTLP_TRACES_CERTIFICATE: "traces/certificate.env",
+            OTEL_EXPORTER_OTLP_TRACES_CLIENT_CERTIFICATE: "traces/client-cert.pem",
+            OTEL_EXPORTER_OTLP_TRACES_CLIENT_KEY: "traces/client-key.pem",
             OTEL_EXPORTER_OTLP_TRACES_COMPRESSION: Compression.Deflate.value,
             OTEL_EXPORTER_OTLP_TRACES_ENDPOINT: "https://traces.endpoint.env",
             OTEL_EXPORTER_OTLP_TRACES_HEADERS: "tracesEnv1=val1,tracesEnv2=val2,traceEnv3===val3==",
@@ -92,6 +104,10 @@ class TestOTLPSpanExporter(unittest.TestCase):
 
         self.assertEqual(exporter._endpoint, "https://traces.endpoint.env")
         self.assertEqual(exporter._certificate_file, "traces/certificate.env")
+        self.assertEqual(
+            exporter._client_certificate_file, "traces/client-cert.pem"
+        )
+        self.assertEqual(exporter._client_key_file, "traces/client-key.pem")
         self.assertEqual(exporter._timeout, 40)
         self.assertIs(exporter._compression, Compression.Deflate)
         self.assertEqual(
@@ -108,6 +124,8 @@ class TestOTLPSpanExporter(unittest.TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_ENDPOINT: OS_ENV_ENDPOINT,
             OTEL_EXPORTER_OTLP_TRACES_ENDPOINT: "https://traces.endpoint.env",
@@ -119,6 +137,8 @@ class TestOTLPSpanExporter(unittest.TestCase):
         exporter = OTLPSpanExporter(
             endpoint="example.com/1234",
             certificate_file="path/to/service.crt",
+            client_key_file="path/to/client-key.pem",
+            client_certificate_file="path/to/client-cert.pem",
             headers={"testHeader1": "value1", "testHeader2": "value2"},
             timeout=20,
             compression=Compression.NoCompression,
@@ -127,6 +147,10 @@ class TestOTLPSpanExporter(unittest.TestCase):
 
         self.assertEqual(exporter._endpoint, "example.com/1234")
         self.assertEqual(exporter._certificate_file, "path/to/service.crt")
+        self.assertEqual(
+            exporter._client_certificate_file, "path/to/client-cert.pem"
+        )
+        self.assertEqual(exporter._client_key_file, "path/to/client-key.pem")
         self.assertEqual(exporter._timeout, 20)
         self.assertIs(exporter._compression, Compression.NoCompression)
         self.assertEqual(
@@ -139,6 +163,8 @@ class TestOTLPSpanExporter(unittest.TestCase):
         "os.environ",
         {
             OTEL_EXPORTER_OTLP_CERTIFICATE: OS_ENV_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE: OS_ENV_CLIENT_CERTIFICATE,
+            OTEL_EXPORTER_OTLP_CLIENT_KEY: OS_ENV_CLIENT_KEY,
             OTEL_EXPORTER_OTLP_COMPRESSION: Compression.Gzip.value,
             OTEL_EXPORTER_OTLP_HEADERS: OS_ENV_HEADERS,
             OTEL_EXPORTER_OTLP_TIMEOUT: OS_ENV_TIMEOUT,
@@ -149,6 +175,10 @@ class TestOTLPSpanExporter(unittest.TestCase):
         exporter = OTLPSpanExporter()
 
         self.assertEqual(exporter._certificate_file, OS_ENV_CERTIFICATE)
+        self.assertEqual(
+            exporter._client_certificate_file, OS_ENV_CLIENT_CERTIFICATE
+        )
+        self.assertEqual(exporter._client_key_file, OS_ENV_CLIENT_KEY)
         self.assertEqual(exporter._timeout, int(OS_ENV_TIMEOUT))
         self.assertIs(exporter._compression, Compression.Gzip)
         self.assertEqual(
@@ -197,7 +227,8 @@ class TestOTLPSpanExporter(unittest.TestCase):
                 (
                     "Header format invalid! Header values in environment "
                     "variables must be URL encoded per the OpenTelemetry "
-                    "Protocol Exporter specification: missingValue"
+                    "Protocol Exporter specification or a comma separated "
+                    "list of name=value occurrences: missingValue"
                 ),
             )