+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/url_attributes.py
@@ -40,7 +40,7 @@ URL_ORIGINAL: Final = "url.original"
 """
 Unmodified original URL as seen in the event source.
 Note: In network monitoring, the observed URL may be a full URL, whereas in access logs, the URL is often just represented as a path. This field is meant to represent the URL as it was observed, complete or not.
-    `url.original` might contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case password and username SHOULD NOT be redacted and attribute's value SHOULD remain the same.
+`url.original` might contain credentials passed via URL in form of `https://username:password@www.example.com/`. In such case password and username SHOULD NOT be redacted and attribute's value SHOULD remain the same.
 """
 
 URL_PATH: Final = "url.path"